# agents: optimize span collector message sending