@santigimeno 2025-04-07
For a nicer visualization:
[report40k.html](https://gistcdn.githack.com/santigimeno/6e8129caa74e579b7fa85562103cede4/raw/bff58b5fce94dd159f8b67bc505bb00cb3ab998a/report40k.html) and [report50k.html](https://gistcdn.githack.com/santigimeno/6e8129caa74e579b7fa85562103cede4/raw/bff58b5fce94dd159f8b67bc505bb00cb3ab998a/report50k.html)