@santigimeno 2025-04-07
Also for max throughput we have gains: [reportwrk.html](https://gistcdn.githack.com/santigimeno/6e8129caa74e579b7fa85562103cede4/raw/9ecfd27c8dd611c159165f02f9a6ff339b84c350/reportwrk.html)
```
=== Benchmark Summary ===

Benchmark Type: wrk (maximum throughput)
Old Version: ./nsolid_baseline
New Version: ./nsolid_span_collector
Iterations: 30
Connections: 10
Duration: 60s
Threads: 2

+------------------+-------------+-------------+------------+-----+
| Metric           | Old Version | New Version | Difference | Sig |
+------------------+-------------+-------------+------------+-----+
| Avg Latency (ms) | 0.65        | 0.64        | N/A        |     |
| P50 Latency (ms) | 0.17        | 0.17        | -1.23%     | **  |
| P90 Latency (ms) | 0.32        | 0.32        | -0.13%     |     |
| P99 Latency (ms) | 0.36        | 0.36        | -1.15%     | *** |
| Avg Req/Sec      | 52,423.01   | 52,955.12   | +1.02%     | **  |
+------------------+-------------+-------------+------------+-----+
```